projects/koenigsegg/.kon/requirements.txt
fastapi[standard]
python-dotenv
sqlmodel
psycopg2-binary
